+++ modules/apps/foundation/portal-configuration/portal-configuration-extender/src/main/java/com/liferay/portal/configuration/extender/internal/NamedConfigurationPathContentFactory.java
@@ -14,10 +14,6 @@
 
 package com.liferay.portal.configuration.extender.internal;
 
-import com.liferay.portal.configuration.extender.BundleStorage;
-import com.liferay.portal.configuration.extender.NamedConfigurationContent;
-import com.liferay.portal.configuration.extender.NamedConfigurationContentFactory;
-import com.liferay.portal.configuration.extender.PropertiesFileNamedConfigurationContent;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.MappingEnumeration;
 import com.liferay.portal.kernel.util.MappingEnumeration.Mapper;
