+++ modules/apps/foundation/portal-configuration/portal-configuration-extender/src/test/java/com/liferay/portal/configuration/extender/internal/NamedConfigurationPathContentFactoryTest.java
@@ -14,9 +14,6 @@
 
 package com.liferay.portal.configuration.extender.internal;
 
-import com.liferay.portal.configuration.extender.BundleStorage;
-import com.liferay.portal.configuration.extender.NamedConfigurationContent;
-import com.liferay.portal.configuration.extender.NamedConfigurationContentFactory;
 import com.liferay.portal.kernel.util.StringUtil;
 
 import java.io.File;
